refactor(service): move cli/cli.ts → src/cli.ts
One fewer directory level; imports updated accordingly.

Committed-By-Agent: claude
